# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyticsEventIn": {
        "properties": {
          "event": {
            "enum": [
              "story_share",
              "notification_received",
              "notification_open"
            ],
            "title": "Event",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          }
        },
        "required": [
          "event"
        ],
        "title": "AnalyticsEventIn",
        "type": "object"
      },
      "AnalyticsEventResponse": {
        "properties": {
          "accepted": {
            "default": true,
            "title": "Accepted",
            "type": "boolean"
          }
        },
        "title": "AnalyticsEventResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HomeResponse": {
        "properties": {
          "top_stories": {
            "items": {
              "$ref": "#/components/schemas/StoryOut"
            },
            "title": "Top Stories",
            "type": "array"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/TopicOut"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "title": "HomeResponse",
        "type": "object"
      },
...
      "PersonalizationOut": {
        "description": "\u00a78.4 explainability: present only when the request supplied\npreferences to rank against, and only ever built from signals that\nactually matched (T16/ADR-005).",
        "properties": {
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "why_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Matters"
          }
        },
        "required": [
          "score"
        ],
        "title": "PersonalizationOut",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StoryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "ShareMetaResponse": {
        "properties": {
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "canonical_url"
        ],
        "title": "ShareMetaResponse",
        "type": "object"
      },
      "StoriesListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StoryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "StoriesListResponse",
        "type": "object"
      },
      "StoryOut": {
        "description": "Appendix A's sample content object, shaped for API consumption.",
        "properties": {
          "canonical_slug": {
            "title": "Canonical Slug",
            "type": "string"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "title": "Importance",
            "type": "number"
          },
          "personalization": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonalizationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "sensitivity": {
            "enum": [
              "NONE",
              "IMMIGRATION",
              "LEGAL",
              "FINANCIAL",
              "BREAKING",
              "OBITUARY_ACCUSATION"
            ],
            "title": "Sensitivity",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/StorySourceOut"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "DRAFT",
              "AI_READY",
              "REVIEW_REQUIRED",
              "APPROVED",
              "SCHEDULED",
              "PUBLISHED",
              "UPDATED",
              "RETRACTED",
              "CORRECTION_PENDING",
              "ARCHIVED"
            ],
            "title": "Status",
            "type": "string"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "title": "Topics",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "variants": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StoryVariantOut"
            },
            "propertyNames": {
              "enum": [
                "en",
                "te"
              ]
            },
            "title": "Variants",
            "type": "object"
          }
        },
        "required": [
          "id",
          "canonical_slug",
          "status",
          "sensitivity",
          "importance",
          "updated_at"
        ],
        "title": "StoryOut",
        "type": "object"
      },
      "StorySourceOut": {
        "properties": {
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "StorySourceOut",
        "type": "object"
      },
      "StoryVariantOut": {
        "properties": {
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "language": {
            "enum": [
              "en",
              "te"
            ],
            "title": "Language",
            "type": "string"
          },
          "qa_status": {
            "default": "PENDING",
            "enum": [
              "PENDING",
              "PASSED",
              "FAILED"
            ],
            "title": "Qa Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "why_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Matters"
          }
        },
        "required": [
          "language",
          "headline",
          "summary"
        ],
        "title": "StoryVariantOut",
        "type": "object"
      },
      "TopicDetailResponse": {
        "properties": {
          "stories": {
            "items": {
              "$ref": "#/components/schemas/StoryOut"
            },
            "title": "Stories",
            "type": "array"
          },
          "topic": {
            "$ref": "#/components/schemas/TopicOut"
          }
        },
        "required": [
          "topic"
        ],
        "title": "TopicDetailResponse",
        "type": "object"
      },
      "TopicOut": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "TopicOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Telugu Global API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events": {
      "post": {
        "description": "T17 \u00a79.4 client-emitted events (`story_share`, `notification_received`,\n`notification_open`) \u2014 the ones the server can't observe on its own\n(a user tapping share, or a device receiving/opening a push). No auth:\nbrowsing/sharing works without login (NON_NEGOTIABLES #9), and there's\nno per-user data here beyond whatever the caller puts in `properties`.",
        "operationId": "track_event_v1_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Event",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/home": {
      "get": {
        "operationId": "get_home_v1_home_get",
        "parameters": [
          {
            "in": "query",
            "name": "residence_country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Residence Country"
            }
          },
          {
            "in": "query",
            "name": "residence_region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Residence Region"
            }
          },
          {
            "in": "query",
            "name": "home_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Home State"
            }
          },
          {
            "in": "query",
            "name": "home_city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Home City"
            }
          },
          {
            "in": "query",
            "name": "topics",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topics"
            }
          },
          {
            "in": "query",
            "name": "segment",
            "required": false,
            "schema": {
              "default": "general",
              "enum": [
                "general",
                "international_student",
                "graduate_opt",
                "professional",
                "family_parent",
                "other"
              ],
              "title": "Segment",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Home",
        "tags": [
          "public"
        ]
      }
    },
...
    "/v1/search": {
      "get": {
        "operationId": "search_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/stories": {
      "get": {
        "operationId": "list_stories_v1_stories_get",
        "parameters": [
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoriesListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Stories",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/stories/{slug}": {
      "get": {
        "operationId": "get_story_v1_stories__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Story",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/stories/{slug}/share-meta": {
      "get": {
        "operationId": "get_story_share_meta_v1_stories__slug__share_meta_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareMetaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Story Share Meta",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/topics/{slug}": {
      "get": {
        "operationId": "get_topic_v1_topics__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Topic",
        "tags": [
          "public"
        ]
      }
    }
  }
}